DaysBack/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
Pillow==9.5.0
psycopg2==2.9.6
pygame==2.4.0
PyJWT==2.6.0
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
